de-stringify and remove circular reference
